Fix: numpy.float64 → Python float 변환으로 PostgreSQL 오류 해결
compute_returns에서 pandas 연산 결과가 numpy.float64 타입으로 반환되어
PostgreSQL이 이를 스키마 참조로 오해하는 문제 수정.
SQLite에서는 타입 무시로 잘 동작했으나 PostgreSQL에서 실패.

+++ backend/app/services/etf_data_fetcher.py
@@ -159,7 +159,7 @@ def compute_returns(tickers: list[str]) -> dict[str, dict]:
                         idx = series.index[series.index >= now - pd.Timedelta(days=days)]
                         if len(idx) > 0:
                             p = series.loc[idx[0]]
-                            r[key] = round(((current - p) / p) * 100, 2)
+                            r[key] = round(float(((current - p) / p) * 100), 2)
 
                     if r:
                         all_results[t] = r
